Add files via upload
"Cams" (short for "cameras") are windows that activly capture and track objects. 
Cams persist, so when an object disappears due to falling below the detection threshold, they can be tracked if and when they reapper. 
However, if they are gone for more than a max still count, the cam is released and available to track a new object.
Image is blurred then a threshold is applied to binary quantize objects, then a threshold is applied.
Tracking ID's are NOT saved